requirements.txt
flask==2.3.3
python-dotenv==1.0.0
openai==1.3.0
supabase==2.0.0
psycopg2-binary==2.9.7
requests==2.31.0